motuscar/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==4.2.21
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-safedelete==1.4.1
djangorestframework==3.16.0
packaging==25.0
pillow==11.2.1
pip==25.1
setuptools==78.1.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
wheel==0.45.1
